Phase 4 batch 2: extract BuildProgressPanel, ShotLightbox, buildActivityFeed
Continues the SubmissionStatusView.tsx decomposition (1,956 -> 1,599 lines).
Three pieces move out, each a genuine seam rather than an arbitrary cut:

- buildActivityFeed.ts: pure logic, no JSX. ActivityEntry, mediaEntries,
  and buildActivityFeed were already shared by three call sites (the main
  component's unread-badge count, ThreadStream's prop type, and
  BuildProgressPanel) — a real shared module, not a relocation for its
  own sake.
- ShotLightbox.tsx: the full-size picture viewer, used by both ThreadStream
  (stays in SubmissionStatusView.tsx) and BuildProgressPanel. Extracted to
  its own file specifically so both can import it without a circular
  dependency between the two panel files.
- BuildProgressPanel.tsx: the checklist + activity-feed render, self-
  contained aside from the two shared pieces above.

All three new files are under the 500-line cap and carry zero comment-prose
debt from the start (short // lines only), rather than repeating last
batch's reseal mistake. No behavior change: all 83 SubmissionStatusView
tests and the full 1,889-test web suite pass unchanged; build is clean.
